@utils/getDroppedFiles.js - one more refactor
